[bugfix] Support Scout in a mismatching Arm binary
When compiling the GOT enable the project to increment the addressess
on their side (thumb host), and also allow for a flag of "is_host_thumb"
for us to do it on their behalf.

This is so we would support a case in which Scout is Thumb but the host
is Arm, or vice versa.